+++ app/src/interfaces/Institution.ts
@@ -1,0 +1,6 @@
+export interface Institution {
+  _id?: string;
+  institutionName: string;
+  domain: string;
+  secondaryDomain: string;
+}
